+++ src/Global.ml
@@ -92,71 +92,71 @@ let print_header_string () =
 
 
 (****************************************************************)
-(** Debug modes *)
+(** Verbose modes *)
 (****************************************************************)
 
-type debug_mode =
-	| Debug_nodebug
-	| Debug_standard
-	| Debug_low
-	| Debug_medium
-	| Debug_high
-	| Debug_total
+type verbose_mode =
+	| Verbose_nodebug
+	| Verbose_standard
+	| Verbose_low
+	| Verbose_medium
+	| Verbose_high
+	| Verbose_total
 
-(* Associate an integer to each debug mode *)
-let level_of_debug = function
-	| Debug_nodebug -> 0
-	| Debug_standard -> 1
-	| Debug_low -> 2
-	| Debug_medium -> 3
-	| Debug_high -> 4
-	| Debug_total -> 5
+(* Associate an integer to each verbose mode *)
+let level_of_verbose = function
+	| Verbose_nodebug -> 0
+	| Verbose_standard -> 1
+	| Verbose_low -> 2
+	| Verbose_medium -> 3
+	| Verbose_high -> 4
+	| Verbose_total -> 5
 
-(* The global debug mode *)
-type global_debug_mode_type =
-	| Debug_mode_not_set
-	| Debug_mode_set of debug_mode
+(* The global verbose mode *)
+type global_verbose_mode_type =
+	| Verbose_mode_not_set
+	| Verbose_mode_set of verbose_mode
 
 (* set to nodebug by default *)
-let global_debug_mode = ref (Debug_mode_set Debug_nodebug)
+let global_verbose_mode = ref (Verbose_mode_set Verbose_nodebug)
 
 let timed_mode = ref false
 
 
-(* Return true if the global debug mode is greater than 'debug_mode', false otherwise *)
-let debug_mode_greater debug_mode =
-	(* Get the global debug mode *)
-	let global_debug_mode =
-	match !global_debug_mode with
-	| Debug_mode_not_set -> raise (InternalError ("The debug mode has not been set, impossible to access it."))
-	| Debug_mode_set global_debug_mode -> global_debug_mode
+(* Return true if the global verbose mode is greater than 'verbose_mode', false otherwise *)
+let verbose_mode_greater verbose_mode =
+	(* Get the global verbose mode *)
+	let global_verbose_mode =
+	match !global_verbose_mode with
+	| Verbose_mode_not_set -> raise (InternalError ("The verbose mode has not been set, impossible to access it."))
+	| Verbose_mode_set global_verbose_mode -> global_verbose_mode
 	in
 	(* Compare *)
-	(level_of_debug global_debug_mode) >= (level_of_debug debug_mode)
+	(level_of_verbose global_verbose_mode) >= (level_of_verbose verbose_mode)
 
 
-(* Convert a string into a debug_mode; raise Not_found if not found *)
-let debug_mode_of_string debug_mode =
-	if debug_mode = "nodebug" then Debug_nodebug
-	else if debug_mode = "standard" then Debug_standard
-	else if debug_mode = "low" then Debug_low
-	else if debug_mode = "medium" then Debug_medium
-	else if debug_mode = "high" then Debug_high
-	else if debug_mode = "total" then Debug_total
+(* Convert a string into a verbose_mode; raise Not_found if not found *)
+let verbose_mode_of_string verbose_mode =
+	if verbose_mode = "nodebug" then Verbose_nodebug
+	else if verbose_mode = "standard" then Verbose_standard
+	else if verbose_mode = "low" then Verbose_low
+	else if verbose_mode = "medium" then Verbose_medium
+	else if verbose_mode = "high" then Verbose_high
+	else if verbose_mode = "total" then Verbose_total
 	else raise Not_found
 
-(* Set the debug mode *)
-let set_debug_mode debug_mode =
-	(*match !global_debug_mode with
-	| Debug_mode_not_set -> global_debug_mode := Debug_mode_set debug_mode
-	| Debug_mode_set debug_mode -> raise (InternalError ("The debug mode has already been set, impossible to set it again."))*)
-	global_debug_mode := Debug_mode_set debug_mode
+(* Set the verbose mode *)
+let set_verbose_mode verbose_mode =
+	(*match !global_verbose_mode with
+	| Verbose_mode_not_set -> global_verbose_mode := Verbose_mode_set verbose_mode
+	| Verbose_mode_set verbose_mode -> raise (InternalError ("The verbose mode has already been set, impossible to set it again."))*)
+	global_verbose_mode := Verbose_mode_set verbose_mode
 
-(* Get the debug mode *)
-let get_debug_mode () =
-	match !global_debug_mode with
-	| Debug_mode_not_set -> raise (InternalError ("The debug mode has not yet been set."))
-	| Debug_mode_set debug_mode -> debug_mode
+(* Get the verbose mode *)
+let get_verbose_mode () =
+	match !global_verbose_mode with
+	| Verbose_mode_not_set -> raise (InternalError ("The verbose mode has not yet been set."))
+	| Verbose_mode_set verbose_mode -> verbose_mode
 
 
 
@@ -388,15 +388,15 @@ let print_message_generic message =
 	flush stdout
 
 
-(* Print a message if global_debug_mode >= message_debug_mode *)
-let print_message message_debug_mode message =
-	(* Only print the message if its message_debug_mode is smaller or equal to the global_debug_mode *)
-	if debug_mode_greater message_debug_mode then
-		(* Compute the debug level *)
-		let debug_level = level_of_debug message_debug_mode in
+(* Print a message if global_verbose_mode >= message_verbose_mode *)
+let print_message message_verbose_mode message =
+	(* Only print the message if its message_verbose_mode is smaller or equal to the global_verbose_mode *)
+	if verbose_mode_greater message_verbose_mode then
+		(* Compute the verbose level *)
+		let verbose_level = level_of_verbose message_verbose_mode in
 		(* Find number of blanks for indentation *)
-		let nb_spaces = if debug_level-1 > 0 then debug_level-1 else 0 in
-		(* Create blanks proportionnally to the debug_level (at least one space) *)
+		let nb_spaces = if verbose_level-1 > 0 then verbose_level-1 else 0 in
+		(* Create blanks proportionnally to the verbose_level (at least one space) *)
 		let spaces = " " ^ string_of_array_of_string (Array.make nb_spaces "   ") in
 		(* Add new lines and blanks everywhere *)
 		let formatted_message = spaces ^ (Str.global_replace (Str.regexp "\n") ("\n" ^ spaces) message) in
@@ -437,7 +437,7 @@ let abort_program () =
 (* Terminate program *)
 let terminate_program () =
 	print_newline();
-	print_message Debug_standard ("PolyOp successfully terminated (" ^ (after_seconds ()) ^ ")");
+	print_message Verbose_standard ("PolyOp successfully terminated (" ^ (after_seconds ()) ^ ")");
 	print_newline();
 	flush stdout;
 	exit(0)